chore: move editor components to sub dir
